finishing the particle effect for the rocket engine "fire"
The particle effect isn't properly generalized for all cases yet, a further study maybe required to keep the glsl particle effect more general.

Motive: For completing the first game (flipbitd), it was decided to rush the particle system a little bit, meaning that the particle component is more specific for the purpose of completing the game rather than allowing the dev to efficienlty use it. Id est, I'm pausing development for now