rm custom iterator from ECS architecture
I just notice this arbitrarly... the iterator solution was really slow. The main problem is that it was using many bound checking, and that it wasn't properly set to be used with SIMD instructions. This new approach is much more manual, but also less abstract, which will I think will be beneficial for the developer. It also is a much more 'pure' approach since it doesn't encapsulate behind a custom iterator.
Lastly, this transparency will make it easier to parallelize later.